requirements.txt
fastapi
uvicorn
pymongo
python-multipart
sqlalchemy
psycopg2-binary
redis
python-jose
passlib